疑似命令を別ソースに移動
