+++ test/CMakeLists.txt
@@ -33,7 +33,8 @@ set(nask_parse_bnfc_SRCS
 
 set(nask_parse_front_SRCS
   ../src/front_end/front_parse.cc
-  ../src/front_end/front_inst.cc)
+  ../src/front_end/front_inst.cc
+  ../src/front_end/front_inst_pseudo.cc)
 
 set(nask_parse_SRCS
   ${nask_parse_bnfc_SRCS}
